Test parent class directly
Although DoctrineORMQuerySourceIteratorTest is named after
DoctrineORMQuerySourceIterator, it only tests getValue(), which is
inherited from AbstractPropertySourceIterator.
We can easily test that method without using
DoctrineORMQuerySourceIterator, but we still need to extend that class
because getValue() is protected, which means we have to make it public
to test it.
As a bonus, we stop extending the class DoctrineORMQuerySourceIterator,
which will be final in next major.
This allows us to avoid a warning from PHPUnit.